% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sua_import.R
\name{sua_import}
\alias{sua_import}
\title{Import FAA Special Use Airspace Order text}
\usage{
sua_import(
  filepath = "http://www.faa.gov/documentLibrary/media/Order/7400.10c_2021-02-09.pdf"
)
}
\arguments{
\item{filepath}{A URL or full filepath to a .txt or .pdf SUA file}
}
\value{
A character vector of text from a SUA order
}
\description{
Import FAA Special Use Airspace Order text
}
